raise play/stop reel button to HomeHeader
