+++ src/components/Home/HomeHeader/HomeHeader.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React, { useState, useEffect, useRef } from "react";
 
 import { AllowCookies, LineBreakerSelector, Nav } from "../../shared/index";
 import { VideoHeader } from "./VideoHeader/VideoHeader";
@@ -8,12 +8,65 @@ import { IHomeHeader } from "../../../interfaces/cms";
 
 import "./HomeHeader.styles.scss";
 
+const controlTextOptions = { play: "Play", stop: "Stop" };
+
 export const HomeHeader = (headerData: IHomeHeader) => {
   const [isPlaying, setIsPlaying] = useState<boolean>(false);
-  const [isNavVisible, setisNavVisible] = useState<boolean>(true);
+  const [isNavVisible, setIsNavVisible] = useState<boolean>(true);
   const [isBurgerMenuOpen, setIsBurgerMenuOpen] = useState<boolean>(false);
+  const [controlText, setControlText] = useState<string>(controlTextOptions.play);
 
-  // TODO: plantear si se queda aquí
+  const videoRef = useRef<HTMLVideoElement>(null);
+  const controlRef = useRef<HTMLDivElement>(null);
+
+  const switchPlayPause = () => {
+    console.log("PLAYPAUSE");
+    setIsPlaying(!isPlaying);
+
+    if (videoRef.current && isPlaying) {
+      videoRef.current.pause();
+      setIsNavVisible(!isNavVisible);
+      setControlText("Play");
+    }
+    if (videoRef.current && !isPlaying) {
+      videoRef.current.play();
+      setControlText("Stop");
+      setTimeout(() => {
+        setIsNavVisible(!isNavVisible);
+      }, 1000);
+    }
+  };
+
+  useEffect(() => {
+    const handleMouseMove = (event: any): void => {
+      if (controlRef.current) {
+        if (event.clientY > 70 && event.clientX < window.innerWidth - 120) {
+          const scrollY = window.scrollY;
+          const postY = event.clientY;
+          const scrollFinalY = scrollY + postY - 10;
+          const scrollX = window.scrollX;
+          const postX = event.clientX;
+          const scrollFinalX = scrollX + postX - 50;
+
+          if (window.innerWidth >= 1040) {
+            controlRef.current.style.top = scrollFinalY.toString().concat("px");
+            controlRef.current.style.left = scrollFinalX.toString().concat("px");
+            controlRef.current.style.opacity = "1";
+          } else {
+            controlRef.current.style.top = (window.innerHeight - 40).toString().concat("px");
+            controlRef.current.style.left = "5%";
+          }
+        } else {
+          controlRef.current.style.opacity = "0";
+        }
+      }
+    };
+    window.addEventListener("mousemove", handleMouseMove);
+    return () => {
+      window.removeEventListener("mousemove", handleMouseMove);
+    };
+  }, []);
+
   function appHeight(): void {
     const doc = document.documentElement;
     doc.style.setProperty("--app-height", `${window.innerHeight}px`);
@@ -35,9 +88,34 @@ export const HomeHeader = (headerData: IHomeHeader) => {
             activeStyle="active-black"
           />
         </span>
-        <div className={isPlaying ? "chazz-title-out" : "chazz-title"}>
-          <LineBreakerSelector typedLines={headerData.title} />
-          <h4>{headerData.subtitle}</h4>
+        <div className="mobile-adjustment">
+          <div className={isPlaying ? "chazz-title-out" : "chazz-title"}>
+            <LineBreakerSelector typedLines={headerData.title} />
+            <h4>{headerData.subtitle}</h4>
+          </div>
+          {!isBurgerMenuOpen && (
+            <>
+              <div className="player-video-mobile-switcher">
+                <div className="player-video">
+                  <div className={`play-icon-${isPlaying ? "in" : "out"}`} onClick={switchPlayPause} />
+                  <div className={`stop-icon-${isPlaying ? "in" : "out"}`} onClick={switchPlayPause} />
+                  <span className="player-text" onClick={switchPlayPause}>
+                    {controlText} reel
+                  </span>
+                </div>
+              </div>
+
+              <div className="player-video-desktop-switcher">
+                <div className="player-video" ref={controlRef}>
+                  <div className={`play-icon-${isPlaying ? "in" : "out"}`} onClick={switchPlayPause} />
+                  <div className={`stop-icon-${isPlaying ? "in" : "out"}`} onClick={switchPlayPause} />
+                  <span className="player-text" onClick={switchPlayPause}>
+                    {controlText} reel
+                  </span>
+                </div>
+              </div>
+            </>
+          )}
         </div>
       </div>
 
@@ -47,8 +125,12 @@ export const HomeHeader = (headerData: IHomeHeader) => {
         isPlaying={isPlaying}
         setIsPlaying={setIsPlaying}
         isNavVisible={isNavVisible}
-        setIsNavVisible={setisNavVisible}
+        setIsNavVisible={setIsNavVisible}
         isBurgerMenuOpen={isBurgerMenuOpen}
+        controlTextOptions={controlTextOptions}
+        controlText={controlText}
+        setControlText={setControlText}
+        ref={videoRef}
       ></VideoHeader>
 
       <AllowCookies />
