+++ src/components/Home/HomeHeader/VideoHeader/VideoHeader.style.scss
@@ -27,9 +27,9 @@
 
 .player-video-mobile-switcher {
   display: inline-flex !important;
-  position: absolute;
-  top: 90%;
-  left: 5%;
+  // position: absolute;
+  // top: 90%;
+  // left: 5%;
   width: 50%;
   cursor: pointer;
 }
