+++ src/components/Home/HomeHeader/VideoHeader/VideoHeader.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useRef, useState } from "react";
+import React, { useEffect, useRef, forwardRef } from "react";
 import "./VideoHeader.style.scss";
 
 type Props = {
@@ -7,14 +7,19 @@ type Props = {
   isNavVisible: boolean;
   setIsNavVisible: (a: boolean) => void;
   isBurgerMenuOpen: boolean;
+  controlTextOptions: { play: string; stop: string };
+  controlText: string;
+  setControlText: (a: string) => void;
 };
 
-const ControlTextOptions = { play: "Play", stop: "Stop" };
+export const VideoHeader = forwardRef(({
+  isPlaying,
+  setIsPlaying,
+  isNavVisible,
+  setIsNavVisible,
+  controlTextOptions,
+  setControlText }: Props, ref: React.LegacyRef<HTMLVideoElement>) => {
 
-export const VideoHeader = ({ isPlaying, setIsPlaying, isNavVisible, setIsNavVisible, isBurgerMenuOpen }: Props) => {
-  const videoRef = useRef<HTMLVideoElement>(null);
-  const controlRef = useRef<HTMLDivElement>(null);
-  const [controlText, setControlText] = useState<string>(ControlTextOptions.play);
 
   useEffect(() => {
     if (!isPlaying && !isNavVisible) {
@@ -22,94 +27,27 @@ export const VideoHeader = ({ isPlaying, setIsPlaying, isNavVisible, setIsNavVis
     }
   }, [isNavVisible, isPlaying, setIsNavVisible]);
 
-  const switchPlayPause = () => {
-    setIsPlaying(!isPlaying);
-
-    if (videoRef.current && isPlaying) {
-      videoRef.current.pause();
-      setIsNavVisible(!isNavVisible);
-      setControlText("Play");
-    }
-    if (videoRef.current && !isPlaying) {
-      videoRef.current.play();
-      setControlText("Stop");
-      setTimeout(() => {
-        setIsNavVisible(!isNavVisible);
-      }, 1000);
-    }
-  };
-
   function forcePause(): void {
     setIsPlaying(false);
-    setControlText(ControlTextOptions.play);
+    setControlText(controlTextOptions.play);
   }
 
   const resetVideo = () => {
     setIsPlaying(false);
     setControlText("Play");
-    if (videoRef.current) {
-      videoRef.current.currentTime = 0;
+    if (ref) {
+      const newRef = ref as React.RefObject<HTMLVideoElement>
+      if (newRef.current) {
+        newRef.current.currentTime = 0;
+      }
     }
   };
 
-  useEffect(() => {
-    const handleMouseMove = (event: any): void => {
-      if (controlRef.current) {
-        if (event.clientY > 70 && event.clientX < window.innerWidth - 120) {
-          const scrollY = window.scrollY;
-          const postY = event.clientY;
-          const scrollFinalY = scrollY + postY - 10;
-          const scrollX = window.scrollX;
-          const postX = event.clientX;
-          const scrollFinalX = scrollX + postX - 50;
-
-          if (window.innerWidth >= 1040) {
-            controlRef.current.style.top = scrollFinalY.toString().concat("px");
-            controlRef.current.style.left = scrollFinalX.toString().concat("px");
-            controlRef.current.style.opacity = "1";
-          } else {
-            controlRef.current.style.top = (window.innerHeight - 40).toString().concat("px");
-            controlRef.current.style.left = "5%";
-          }
-        } else {
-          controlRef.current.style.opacity = "0";
-        }
-      }
-    };
-    window.addEventListener("mousemove", handleMouseMove);
-    return () => {
-      window.removeEventListener("mousemove", handleMouseMove);
-    };
-  }, []);
-
   return (
     <>
-      {!isBurgerMenuOpen && (
-        <>
-          <div className="player-video-mobile-switcher">
-            <div className="player-video">
-              <div className={`play-icon-${isPlaying ? "in" : "out"}`} onClick={switchPlayPause} />
-              <div className={`stop-icon-${isPlaying ? "in" : "out"}`} onClick={switchPlayPause} />
-              <span className="player-text" onClick={switchPlayPause}>
-                {controlText} reel
-              </span>
-            </div>
-          </div>
-
-          <div className="player-video-desktop-switcher">
-            <div className="player-video" ref={controlRef}>
-              <div className={`play-icon-${isPlaying ? "in" : "out"}`} onClick={switchPlayPause} />
-              <div className={`stop-icon-${isPlaying ? "in" : "out"}`} onClick={switchPlayPause} />
-              <span className="player-text" onClick={switchPlayPause}>
-                {controlText} reel
-              </span>
-            </div>
-          </div>
-        </>
-      )}
       <video
         className={`video-header ${isPlaying ? "color" : "no-color"}`}
-        ref={videoRef}
+        ref={ref}
         onEnded={resetVideo}
         onPause={forcePause}
         preload="auto"
@@ -119,4 +57,4 @@ export const VideoHeader = ({ isPlaying, setIsPlaying, isNavVisible, setIsNavVis
       </video>
     </>
   );
-};
+});
